Integration Market and Other

+++ Assets/Script/Gameplay/Config/CharacterStarterConfigSO.cs
@@ -9,13 +9,13 @@ namespace SotongStudio.Bomber.Gameplay.Character
         [SerializeField]
         private int _maxHealth;
         [SerializeField]
-        private int _speed;
+        private float _speed;
         [SerializeField]
         private int _maxBombAmount;
 
         public int MaxHealth => _maxHealth;
         public int Health => _maxHealth;
-        public int Speed => _speed;
+        public float Speed => _speed;
         public int BombAmount => _maxBombAmount;
         public int MaxBombAmount => _maxBombAmount;
